Build: pacemakerd: rename source files
